fix(ui): JourneyForm now shows all steps.
the solution i preseribed previously was the wrong one. the actual
problem was due to the onSubmit nature of the form. in the previous
iteration, when someone got to the final stage in which the button was
of type submit, for some reason when moving to that stage, the form
would auto-submit before allowing a user to change the fields in that
stage. the change was to make the final button of type button, and make
the onClick method call handleSubmit, instead of the <Form /> element
calling submit from the button of type submit.